requirements.txt
fastapi
uvicorn
pydantic[email]
motor
pymongo
python-dotenv
bcrypt
python-jose[cryptography]
